Interface ofport and ofport_request support
Added support to yang model for ofport and ofport_request.  These fields are
settable at the time of endpoint creation.  Changed the Interface schema ofport
field to utilize the type Integer; ofport is defined as an integer in the
Open_vSwitch specification.  Changed references to Interface ofport column to
use Integer data type.  Enforced a Yang model range for ofport_request to
ensure the value is in the range [1,65279] as defined by the Open_vSwitch
specification. Future work will persist these settings to the data store.

+++ openstack/net-virt-providers/src/main/java/org/opendaylight/ovsdb/openstack/netvirt/providers/openflow13/OF13Provider.java
@@ -854,7 +854,7 @@ public class OF13Provider implements NetworkingProvider {
                 return;
             }
 
-            Set<Long> of_ports = intf.getOpenFlowPortColumn().getData();
+            Set<Integer> of_ports = intf.getOpenFlowPortColumn().getData();
             if (of_ports == null || of_ports.size() <= 0) {
                 logger.debug("Could NOT Identify OF value for port {} on {}", intf.getName(), node);
                 return;
@@ -909,7 +909,7 @@ public class OF13Provider implements NetworkingProvider {
                 return;
             }
 
-            Set<Long> of_ports = intf.getOpenFlowPortColumn().getData();
+            Set<Integer> of_ports = intf.getOpenFlowPortColumn().getData();
             if (of_ports == null || of_ports.size() <= 0) {
                 logger.debug("Could NOT Identify OF value for port {} on {}", intf.getName(), node);
                 return;
@@ -955,7 +955,7 @@ public class OF13Provider implements NetworkingProvider {
                 return;
             }
 
-            Set<Long> of_ports = intf.getOpenFlowPortColumn().getData();
+            Set<Integer> of_ports = intf.getOpenFlowPortColumn().getData();
             if (of_ports == null || of_ports.size() <= 0) {
                 logger.debug("Could NOT Identify OF value for port {} on {}", intf.getName(), node);
                 return;
@@ -1021,7 +1021,7 @@ public class OF13Provider implements NetworkingProvider {
                 return;
             }
 
-            Set<Long> of_ports = intf.getOpenFlowPortColumn().getData();
+            Set<Integer> of_ports = intf.getOpenFlowPortColumn().getData();
             if (of_ports == null || of_ports.size() <= 0) {
                 logger.error("Could NOT Identify OF value for port {} on {}", intf.getName(), node);
                 return;
@@ -1084,7 +1084,7 @@ public class OF13Provider implements NetworkingProvider {
                 return;
             }
 
-            Set<Long> of_ports = intf.getOpenFlowPortColumn().getData();
+            Set<Integer> of_ports = intf.getOpenFlowPortColumn().getData();
             int timeout = 6;
             while ((of_ports == null) && (timeout > 0)) {
                 of_ports = intf.getOpenFlowPortColumn().getData();
@@ -1164,7 +1164,7 @@ public class OF13Provider implements NetworkingProvider {
                 return;
             }
 
-            Set<Long> of_ports = intf.getOpenFlowPortColumn().getData();
+            Set<Integer> of_ports = intf.getOpenFlowPortColumn().getData();
             if (of_ports == null || of_ports.size() <= 0) {
                 logger.error("Could NOT Identify OF value for port {} on {}", intf.getName(), node);
                 return;
